==> code/Step1.1_load_traffic_flow_group_by_road_python.ipynb <==
import pandas as pd
import numpy as np



traffic_flow_hourly = pd.read_csv(r'N:\GIS Bicycle Project\traffic_flow\comptages-routiers-permanents-2019.csv',sep = ';')
traffic_flow_hourly.head()

traffic_flow_hourly.columns

df = traffic_flow_hourly.groupby('iu_ac').mean()
df.head()

df.describe()

df[['q', 'k']].to_csv(r'N:\GIS Bicycle Project\traffic_flow\traffic_flow_by_road.csv', index = True, header=True)
